perf: riscv: use capabilities for saved registers
Some of the perf code uses register values that come from struct pt_regs
or from the compiler's __builtin_return_address. Both of them provide
capabilities on a cheri system.

Update some of the code to use capabilities for saved registers.

+++ arch/riscv/include/asm/perf_event.h
@@ -14,7 +14,7 @@
 
 #define perf_arch_fetch_caller_regs(regs, __ip) { \
 	(regs)->epc = (__ip); \
-	(regs)->s0 = (unsigned long) __builtin_frame_address(0); \
+	(regs)->s0 = (uintptr_t)__builtin_frame_address(0); \
 	(regs)->sp = current_stack_pointer; \
 	(regs)->status = SR_PP; \
 }
